+++ tests/index.test.ts
@@ -198,7 +198,7 @@ describe('instantiate client', () => {
 
   test('with environment variable arguments', () => {
     // set options via env var
-    process.env['FLOWGLAD_API_KEY'] = 'My API Key';
+    process.env['FLOWGLAD_SECRET_KEY'] = 'My API Key';
     const client = new Flowglad();
     expect(client.apiKey).toBe('My API Key');
   });